Created arrayChange.java #17 in Waters Lvl 4
You are given an array of integers. On each move you are allowed to increase exactly one of its element by one. Find the minimal number of moves required to obtain a strictly increasing sequence from the input.

Example

For inputArray = [1, 1, 1], the output should be
arrayChange(inputArray) = 3. (bc [1,1(+1),1(+2)] for inc. order [1,2,3])